+++ modules/Student_Billing/MassAssignFees.php
@@ -11,30 +11,23 @@ if ( $_REQUEST['modfunc'] === 'save' )
 		{
 			$due_date = RequestedDate( 'due', '' );
 
-			if ( $due_date )
+			// Group SQL inserts.
+			$sql = '';
+
+			foreach ( (array) $_REQUEST['student'] as $student_id )
 			{
-				// Group SQL inserts.
-				$sql = '';
-
-				foreach ( (array) $_REQUEST['student'] as $student_id )
-				{
-					$sql .= "INSERT INTO BILLING_FEES (STUDENT_ID,ID,TITLE,AMOUNT,SYEAR,SCHOOL_ID,ASSIGNED_DATE,DUE_DATE,COMMENTS)
-						VALUES('" . $student_id . "'," . db_seq_nextval( 'billing_fees_id_seq' ) . ",
-						'" . $_REQUEST['title'] . "','" . preg_replace( '/[^0-9.-]/', '', $_REQUEST['amount'] ) . "',
-						'" . UserSyear() . "','" . UserSchool() . "','" . DBDate() . "','" . $due_date . "',
-						'" . $_REQUEST['comments'] . "');";
-				}
-
-				if ( $sql )
-				{
-					DBQuery( $sql );
-
-					$note[] = button( 'check' ) . '&nbsp;' . _( 'That fee has been added to the selected students.' );
-				}
+				$sql .= "INSERT INTO BILLING_FEES (STUDENT_ID,ID,TITLE,AMOUNT,SYEAR,SCHOOL_ID,ASSIGNED_DATE,DUE_DATE,COMMENTS)
+					VALUES('" . $student_id . "'," . db_seq_nextval( 'billing_fees_id_seq' ) . ",
+					'" . $_REQUEST['title'] . "','" . preg_replace( '/[^0-9.-]/', '', $_REQUEST['amount'] ) . "',
+					'" . UserSyear() . "','" . UserSchool() . "','" . DBDate() . "','" . $due_date . "',
+					'" . $_REQUEST['comments'] . "');";
 			}
-			else
+
+			if ( $sql )
 			{
-				$error[] = _( 'The date you entered is not valid' );
+				DBQuery( $sql );
+
+				$note[] = button( 'check' ) . '&nbsp;' . _( 'That fee has been added to the selected students.' );
 			}
 		}
 		else
@@ -68,27 +61,34 @@ if ( ! $_REQUEST['modfunc'] )
 
 		PopTable( 'header', _( 'Fee' ) );
 
-		echo '<table class="col1-align-right">';
+		echo '<table><tr><td>' . TextInput(
+			'',
+			'title',
+			_( 'Title' ),
+			'required size="20"'
+		) . '</td></tr>';
 
-		echo '<tr><td>' . _( 'Title' ) . '</td><td><input type="text" name="title" required /></td></tr>';
+		echo '<tr><td>' . TextInput(
+			'',
+			'amount',
+			_( 'Amount' ),
+			'size="5" maxlength="10" required'
+		) . '</td></tr>';
 
-		echo '<tr><td>' . _( 'Amount' ) . '</td><td><input type="text" name="amount" size="5" maxlength="10" required /></td></tr>';
+		echo '<tr><td>' . DateInput( '', 'due', _( 'Due Date' ), false ) . '</td></tr>';
 
-		echo '<tr><td>' . _( 'Due Date' ) . '</td>
-			<td>' . DateInput( DBDate(), 'due', '', false, false ) . '</td></tr>';
-
-		echo '<tr><td>' . _( 'Comment' ) . '</td><td><input type="text" name="comments" /></td></tr>';
-
-		echo '</table>';
+		echo '<tr><td>' . TextInput(
+			'',
+			'comments',
+			_( 'Comment' ),
+			'maxlength="1000" size="25"'
+		) . '</td></tr></table>';
 
 		PopTable( 'footer' );
 
 		echo '<br />';
 	}
-}
 
-if ( ! $_REQUEST['modfunc'] )
-{
 	$extra['link'] = array( 'FULL_NAME' => false );
 	$extra['SELECT'] = ",NULL AS CHECKBOX";
 	$extra['functions'] = array( 'CHECKBOX' => 'MakeChooseCheckbox' );